requirements.txt
fastapi==0.71.0
firebase-admin==5.2.0
pip == 21.3.1
tweepy==4.4.0
uvicorn==0.15.0
uvloop==0.16.0
python-dotenv==0.19.2
schedule==1.1.0
emoji==1.6.3
neologdn==0.5.1
janome==0.4.1